+++ package.json
@@ -138,11 +138,6 @@
       "import": "./dist/middleware/jsx-renderer/index.js",
       "require": "./dist/cjs/middleware/jsx-renderer/index.js"
     },
-    "./jsx/hooks": {
-      "types": "./dist/types/jsx/hooks/index.d.ts",
-      "import": "./dist/jsx/hooks/index.js",
-      "require": "./dist/cjs/jsx/hooks/index.js"
-    },
     "./jsx/dom": {
       "types": "./dist/types/jsx/dom/index.d.ts",
       "import": "./dist/jsx/dom/index.js",
@@ -370,9 +365,6 @@
       ],
       "jsx-renderer": [
         "./dist/types/middleware/jsx-renderer"
-      ],
-      "jsx/hooks": [
-        "./dist/types/jsx/hooks"
       ],
       "jsx/dom": [
         "./dist/types/jsx/dom"
