+++ src/jsx/index.ts
@@ -4,6 +4,17 @@ import type { StringBuffer, HtmlEscaped, HtmlEscapedString } from '../utils/html
 import type { IntrinsicElements as IntrinsicElementsDefined } from './intrinsic-elements'
 export { ErrorBoundary } from './components'
 export { Suspense } from './streaming'
+export {
+  useState,
+  useEffect,
+  useRef,
+  useCallback,
+  use,
+  startTransition,
+  useTransition,
+  useDeferredValue,
+} from './hooks'
+export type { RefObject } from './hooks'
 
 export const HONO_COMPONENT = 'hono-component'
 export const HONO_COMPONENT_ID = 'hono-component-id'
